Factor out get_stats functions
