+++ mica/stats/update_acq_stats.py
@@ -701,22 +701,6 @@ def update(opt):
         _save_acq_stats(t)
 
 
-def get_stats(filter=True):
-    """
-    Retrieve numpy array of acq stats
-
-    :param filter: True filters out 'known_bad' rows from the table
-    :returns acq_stats: numpy.ndarray
-    """
-
-    h5 = tables.open_file(table_file, 'r')
-    stats = h5.root.data[:]
-    h5.close()
-    if filter:
-        stats = stats[~stats['known_bad']]
-    return stats
-
-
 def main():
     opt = get_options()
     if opt.email is None:
