+++ mica/stats/update_guide_stats.py
@@ -548,22 +548,6 @@ def update(opt):
         _save_gui_stats(t)
 
 
-def get_stats(filter=True):
-    """
-    Retrieve numpy array of guide stats
-
-    :param filter: True filters out 'known_bad' rows from the table
-    :returns gui_stats: numpy.ndarray
-    """
-
-    h5 = tables.open_file(TABLE_FILE, 'r')
-    stats = h5.root.data[:]
-    h5.close()
-    if filter:
-        stats = stats[~stats['known_bad']]
-    return stats
-
-
 def main():
     opt = get_options()
     update(opt)
